Fix decoding nil into struct field of type pointer
When decoding nil into a struct field of type pointer, the field value
should be set to nil.  We used to allocate a new default value for the
pointer to point to.  Because of this, decoding can produce a different
object from encoding input object.

Added test to prevent regression.