Add CBOR duplicate map key options (#152)
This CBOR library provides 2 options for duplicate map keys:

- DupMapKeyQuiet: Turn off detection of duplicate map keys. It tries
to use a "keep fastest" method by choosing either "keep first" or
"keep last" depending on the Go data type.

- DupMapKeyEnforcedAPF: Turn on detection and rejection of duplidate
map keys. Decoding stops immediately and returns DupMapKeyError when the
first duplicate key is detected. The error includes the duplicate map key
and the index number.

APF suffix means "Allow Partial Fill" so the destination map or struct
can contain some decoded values at the time of error. It is the caller's
responsibility to respond to the DuplicateMapKeyErr by discarding the
partially filled result if that's required by their protocol.

Detection of duplicate map keys relies on whether the CBOR map key would
be a duplicate "key" when decoded and applied to the user-provided Go map
or struct.