Fix link to badge.svg for code coverage
Unfortunately, the price of a more privacy is having to update README.md after updating ci-go-cover.yml with new minimum required coverage.

Basically, the link to the badge.svg needs to match the name specified in ci-go-cover.yml.  Otherwise, it'll display the last badge.svg generaged by GitHub and also link to the old ci build (those are always in sync.)

After this change, the badge will correctly show 98% (pass/fail) status.  And the new destination URL only displays the CI info for builds matching 98%.  It doesn't any of the 97% builds which kinda makes sense but isn't intuitive.

From:  
 [![](https://github.com/fxamacker/cbor/workflows/cover%20%E2%89%A598%25/badge.svg)](https://github.com/fxamacker/cbor/actions?query=workflow%3A%22cover+%E2%89%A598%25%22)

To:  
 [![](https://github.com/fxamacker/cbor/workflows/cover%20%E2%89%A597%25/badge.svg)](https://github.com/fxamacker/cbor/actions?query=workflow%3A%22cover+%E2%89%A597%25%22)