Strip CBOR tag number 55799 when it is a prefix
Tag number 55799 is normally prefixed to data to indicate that it is
CBOR encoded.

This library strips tag number 55799 (when it is a prefix) to provide
the same decoding behavior as if CBOR data is not tagged.

* When decoding to empty interface or other Go types, tag number 55799
is skipped and tag content is decoded.

* When decoding to Unmarshaler, tag number 55799 is skipped and only
the tag content is provided to UnmarshalCBOR.

Closes #228