Add InvalidMapKeyTypeError
InvalidMapKeyTypeError is returned when attempting
to decode CBOR map key to a Go type that cannot
be used as map key.

This allows callers to check for this error without
requiring string comparison of the error text.

To maintain backward compatibility, the text
of InvalidMapKeyTypeError remains the same
as the plain error that used to be returned.