Fix handling extra data in Unmarshal() & Valid()
Previously, Unmarshal() and Valid() ignored extra
data (if any) after the current CBOR data item. This
matched the behavior of Decoder.Decode().

However, Unmarshal() and Valid() are typically used for
single CBOR data item, so extra data (if present)
should return ExtraneousDataError.

This is a bug fix because the previous behavior
of ignoring extra data didn't exactly match RFC 8949
regarding this and was also different from how
encoding/json handles this.

The behavior for Decoder.Decode() is unchanged because
extra data after reading a single CBOR data item
is expected for Decoder with io.Reader.

Special thanks to zensh for reporting this issue
and proposing a solution.