Remove ignored UTF8 setting in diagMode()
Unreleased functions Diagnose() and DiagnoseFirst() added in v2.5.0-beta3
were ignoring the setting of DecOptions.UTF8 = DecodeInvalidUTF8 in
the underlying decoder.

Given this, removing the ignored code is enough to make the new
diagnostic functions match the default behavior of CBOR decoder
which is to reject invalid UTF-8 in CBOR text strings.

For now, it doesn't make sense to print invalid UTF-8 text
because the Diagnostic Notation described by RFC 8949 and the
Extended Diagnostic Notation in Appendix G of RFC 8610 don't specify
how to represent invalid UTF-8.