Add encoding option to specify how omitempty fields are encoded
The current behavior in the library is to omits if field value
would encode as empty JSON value. This behavior is different from
the bahavior in golang encoding/json which omits if field value
is an empty Go value (defined as false, 0, a nil pointer, a nil
interface value, and any empty array, slice, map, or string).
The OmitEmptyMode = 1 is added to encoding option so that cbor
would decode omitempty fields similarly to encoding/json in go.

Signed-off-by: Vu Dinh <vudinh@outlook.com>