Add option to configure how simple values are decoded.
Applications can use the decode option SimpleValues to override the default unmarshal behavior on a
per-simple-value basis. The option is open to future extension, and currently supports the choice
between the default backwards-compatible behavior (true as true, false as false, null as nil, and
undefined as nil, and cbor.SimpleValue(N) for each currently-unregistered but well-formed simple
value number N) and outright rejection with the new error type UnacceptableDataItemError.

Signed-off-by: Ben Luddy <bluddy@redhat.com>