Improve DupMapKeyError message
This commit is only relevant for user apps that use
the non-default DupMapKeyEnforcedAPF decoding option.

Currently, DupMapKeyError uses hardcoded quotes in the
error message for map keys, including non-string types.

Given this, an error message containing a quoted integer
having a value of 1 looks the same as the string "1"
because they both have quotes in the error message.

This commit removes the hardcoded quotes in the error message
and replaces %v with %#v to only emit quotes when applicable.