chore: export constants for external use
in some cases it is specially useful to know what are the maximum values
allowed by this package. for example here:

https://github.com/kubernetes/kubernetes/pull/135340

we had to copy the values instead of referring to them. this broads the
public scope of this package, if we don't want this then please feel
free to close or suggest another approach.

Signed-off-by: Ricardo Maraschini<ricardo.maraschini@gmail.com>